requirements.txt
streamlit
langchain
mysql-connector-python
python-dotenv